+++ packages/react-learn-card/src/components/VCDisplayCard2/VCDisplayCard2.tsx
@@ -177,7 +177,10 @@ export const VCDisplayCard2: React.FC<VCDisplayCard2Props> = ({
 
     const _title = titleOverride || title;
 
-    if (categoryType === LCCategoryEnum.meritBadge) {
+    if (
+        categoryType === LCCategoryEnum.meritBadge ||
+        credential?.display?.displayType === 'award'
+    ) {
         return (
             <MeritBadgeDisplayCard
                 credential={credential}
@@ -323,10 +326,7 @@ export const VCDisplayCard2: React.FC<VCDisplayCard2Props> = ({
                         </Flipped>
                     )}
 
-                    <div
-                        className="relative vc-card-content-container flex flex-col items-center grow min-h-0 w-full rounded-t-[30px] bg-[#353E64] rounded-b-[200px]"
-                        // style={backgroundStyle}
-                    >
+                    <div className="relative vc-card-content-container flex flex-col items-center grow min-h-0 w-full rounded-t-[30px] bg-[#353E64] rounded-b-[200px]">
                         <Flipped inverseFlipId="card" scale>
                             <div
                                 className="absolute top-0 left-0 w-full h-full rounded-b-[200px] rounded-t-[30px]"
@@ -404,16 +404,16 @@ export const VCDisplayCard2: React.FC<VCDisplayCard2Props> = ({
                                     {isFront && customFrontButton}
                                     {((isFront && !customFrontButton) ||
                                         (isFront && showDetailsBtn)) && (
-                                        <Flipped inverseFlipId="card">
-                                            <button
-                                                type="button"
-                                                className="vc-toggle-side-button text-white shadow-bottom bg-[#00000099] px-[30px] py-[8px] rounded-[40px] text-[28px] tracking-[0.75px] uppercase leading-[28px] mt-[40px] w-fit select-none"
-                                                onClick={() => setIsFront(!isFront)}
-                                            >
-                                                Details
-                                            </button>
-                                        </Flipped>
-                                    )}
+                                            <Flipped inverseFlipId="card">
+                                                <button
+                                                    type="button"
+                                                    className="vc-toggle-side-button text-white shadow-bottom bg-[#00000099] px-[30px] py-[8px] rounded-[40px] text-[28px] tracking-[0.75px] uppercase leading-[28px] mt-[40px] w-fit select-none"
+                                                    onClick={() => setIsFront(!isFront)}
+                                                >
+                                                    Details
+                                                </button>
+                                            </Flipped>
+                                        )}
                                     {!isFront && (
                                         <Flipped inverseFlipId="card">
                                             <button